OneOnOne/requirements.txt
django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
django-environ==0.11.2
django-cors-headers==4.3.1